isolate Page model tests
